Move Checkbox from UI to Onboarding
