+++ apps/onboarding/package.json
@@ -19,6 +19,7 @@
     "@datadog/browser-rum": "4.43.0",
     "@emotion/react": "11.11.1",
     "@emotion/styled": "11.11.0",
+    "@radix-ui/react-checkbox": "1.0.4",
     "@radix-ui/react-radio-group": "1.1.3",
     "classnames": "2.3.2",
     "cookies-next": "2.1.2",
