+++ packages/ui/package.json
@@ -50,8 +50,6 @@
   "dependencies": {
     "@emotion/react": "11.11.1",
     "@emotion/styled": "11.11.0",
-    "@radix-ui/react-checkbox": "1.0.4",
-    "@radix-ui/react-radio-group": "1.1.3",
     "react": "18.2.0",
     "react-dom": "18.2.0"
   }
